+++ pydifact/segmentcollection.py
@@ -21,8 +21,11 @@
 # THE SOFTWARE.
 
 import collections
-from typing import List, Optional
+from typing import List, Optional, Tuple, Union
+import datetime
+import warnings
 
+from pydifact.api import EDISyntaxError
 from pydifact.parser import Parser
 from pydifact.segments import Segment
 from pydifact.serializer import Serializer
@@ -30,40 +33,28 @@ from pydifact.control import Characters
 import codecs
 
 
-class SegmentCollection:
+class AbstractSegmentsContainer:
     """Represent a collection of EDI Segments for both reading and writing."""
 
-    def __init__(self):
+    def __init__(self, extra_header_elements: List[Union[str, List[str]]] = []):
+        """
+        :param extra_header_elements: a list of elements to be appended at the end
+          of the header segment (same format as Segment() constructor *elements).
+        """
 
         # The segments that make up this message
         self.segments = []
         self.characters = Characters()
 
+        self.extra_header_elements = extra_header_elements
+
         # Flag whether the UNA header is present
         self.has_una_segment = False
-
-    @classmethod
-    def from_file(cls, file: str, encoding: str = "iso8859-1") -> "SegmentCollection":
-        """Create a SegmentCollection instance from a file.
-
-        Raises FileNotFoundError if filename is not found.
-        :param encoding: an optional string which specifies the encoding. Default is "iso8859-1".
-        :param file: The full path to a file that contains an EDI message.
-        :rtype: SegmentCollection
-        """
-
-        # codecs.lookup raises an LookupError if given codec was not found:
-        codecs.lookup(encoding)
-
-        with open(file, encoding=encoding) as f:
-            collection = f.read()
-        return cls.from_str(collection)
 
     @classmethod
     def from_str(cls, string: str) -> "SegmentCollection":
         """Create a SegmentCollection instance from a string.
         :param string: The EDI content
-        :rtype: SegmentCollection
         """
         segments = Parser().parse(string)
 
@@ -72,15 +63,14 @@ class SegmentCollection:
     @classmethod
     def from_segments(
         cls, segments: list or collections.Iterable
-    ) -> "SegmentCollection":
-        """Create a new SegmentCollection instance from a iterable list of segments.
+    ) -> "AbstractSegmentsContainer":
+        """Create a new AbstractSegmentsContainer instance from a iterable list of segments.
 
         :param segments: The segments of the EDI interchange
         :type segments: list/iterable of Segment
-        :rtype: SegmentCollection
         """
 
-        # create a new instance of SegmentCollection and return it
+        # create a new instance of AbstractSegmentsContainer and return it
         # with the added segments
         return cls().add_segments(segments)
 
@@ -106,7 +96,7 @@ class SegmentCollection:
 
     def add_segments(
         self, segments: List[Segment] or collections.Iterable
-    ) -> "SegmentCollection":
+    ) -> "AbstractSegmentsContainer":
         """Add multiple segments to the collection. Passing a UNA segment means setting/overriding the control
         characters and setting the serializer to output the Service String Advice. If you wish to change the control
         characters from the default and not output the Service String Advice, change self.characters instead,
@@ -120,7 +110,80 @@ class SegmentCollection:
 
         return self
 
-    def add_segment(self, segment: Segment) -> "SegmentCollection":
+    def add_segment(self, segment: Segment) -> "AbstractSegmentsContainer":
+        """Append a segment to the collection.
+
+        :param segment: The segment to add
+        """
+        self.segments.append(segment)
+        return self
+
+    def get_header_segment(self) -> Optional[Segment]:
+        """Craft and return this container header segment (if any)
+
+        :returns: None if there is no header for that container
+        """
+        return None
+
+    def get_footer_segment(self) -> Optional[Segment]:
+        """Craft and return this container footer segment (if any)
+        :returns: None if there is no footer for that container
+        """
+        return None
+
+    def serialize(self, break_lines: bool = False) -> str:
+        """Serialize all the segments added to this object.
+        :param break_lines: if True, insert line break after each segment terminator.
+        """
+        header = self.get_header_segment()
+        footer = self.get_footer_segment()
+        out = []
+
+        if header:
+            out.append(header)
+        out += self.segments
+        if footer:
+            out.append(footer)
+
+        return Serializer(self.characters).serialize(
+            out, self.has_una_segment, break_lines,
+        )
+
+    def __str__(self) -> str:
+        """Allow the object to be serialized by casting to a string."""
+        return self.serialize()
+
+
+class FileSourcableMixin:
+    """
+    For backward compatibility
+
+    For v0.2 drop this class and move from_file() to Interchange class.
+    """
+    @classmethod
+    def from_file(cls, file: str, encoding: str = "iso8859-1") -> "FileSourcableMixin":
+        """Create a Interchange instance from a file.
+
+        Raises FileNotFoundError if filename is not found.
+        :param encoding: an optional string which specifies the encoding. Default is "iso8859-1".
+        :param file: The full path to a file that contains an EDI message.
+        :rtype: FileSourcableMixin
+        """
+        # codecs.lookup raises an LookupError if given codec was not found:
+        codecs.lookup(encoding)
+
+        with open(file, encoding=encoding) as f:
+            collection = f.read()
+        return cls.from_str(collection)
+
+class UNAHandlingMixin:
+    """
+    For backward compatibility
+
+    For v0.2 drop this class and move add_segment() to Interchange class.
+    """
+
+    def add_segment(self, segment: Segment) -> "UNAHandlingMixin":
         """Append a segment to the collection. Passing a UNA segment means setting/overriding the control
         characters and setting the serializer to output the Service String Advice. If you wish to change the control
         characters from the default and not output the Service String Advice, change self.characters instead,
@@ -132,17 +195,218 @@ class SegmentCollection:
             self.has_una_segment = True
             self.characters = Characters.from_str(segment.elements[0])
             return self
-        self.segments.append(segment)
-        return self
+        return super().add_segment(segment)
 
-    def serialize(self, break_lines: bool = False) -> str:
-        """Serialize all the segments added to this object.
-        :param break_lines: if True, insert line break after each segment terminator.
+
+
+class SegmentCollection(FileSourcableMixin, UNAHandlingMixin, AbstractSegmentsContainer):
+    """
+    For backward compatibility. Drop it in v0.2
+
+    Will be replaced by Interchange or RawSegmentCollection depending on the need.
+    """
+    def __init__(self, *args, **kwargs):
+        warnings.warn(
+            "SegmentCollection is deprecated and will no longer be available in v0.2, "
+            "replace it with Interchange or RawSegmentCollection",
+            DeprecationWarning,
+        )
+        super().__init__(*args, **kwargs)
+
+    @classmethod
+    def from_file(cls, *args, **kwargs) -> "SegmentCollection":
+        warnings.warn(
+            "SegmentCollection.from_file will be removed in v0.2, "
+            "Use Interchange class instead",
+            DeprecationWarning,
+        )
+        return  super().from_file(*args, **kwargs)
+
+    def add_segment(self, segment: Segment) -> "SegmentCollection":
+        if segment.tag == "UNA":
+            warnings.warn(
+                "SegmentCollection will be removed in v0.2, "
+                "For UNA handling, use Interchange class instead",
+                DeprecationWarning,
+            )
+        return super().add_segment(segment)
+
+
+class RawSegmentCollection(AbstractSegmentsContainer):
+    """
+    A way to analyze arbitrary bunch of edifact segments.
+
+    Similar to the deprecated SegmentCollection, but lacking from_file() and UNA support.
+
+    If you are handling an Interchange or a Message, you may want to prefer
+    those classes to RawSegmentCollection, as they offer more features and
+    checks.
+    """
+    pass
+
+
+
+class Message(AbstractSegmentsContainer):
+    """
+    A message (started by UNH segment, ended by UNT segment)
+
+    Optional features of UNH are not yet supported.
+
+    https://www.stylusstudio.com/edifact/40100/UNH_.htm
+    https://www.stylusstudio.com/edifact/40100/UNT_.htm
+    """
+    def __init__(
+            self,
+            reference_number: str,
+            identifier: Tuple,
+            *args, **kwargs
+    ):
+        super().__init__(*args, **kwargs)
+        self.reference_number = reference_number
+        self.identifier = identifier
+
+    @property
+    def type(self) -> str:
+        return self.identifier[0]
+
+    @property
+    def version(self) -> str:
         """
-        return Serializer(self.characters).serialize(
-            self.segments, self.has_una_segment, break_lines
+        Gives version number and release number.
+
+        :return: message version, parsable by pkg_resources.parse_version()
+        """
+        return f'{self.identifier[1]}.{self.identifier[2]}'
+
+
+    def get_header_segment(self) -> Segment:
+        return Segment(
+            "UNH",
+            self.reference_number,
+            [str(i) for i in self.identifier],
+            *self.extra_header_elements,
         )
 
-    def __str__(self) -> str:
-        """Allow the object to be serialized by casting to a string."""
-        return self.serialize()
+    def get_footer_segment(self) -> Segment:
+        return Segment(
+            "UNT",
+            self.reference_number,
+            str(len(self.segments)),
+        )
+
+
+class Interchange(FileSourcableMixin, UNAHandlingMixin, AbstractSegmentsContainer):
+    """
+    An interchange (started by UNB segment, ended by UNZ segment)
+
+    Optional features of UNB are not yet supported.
+
+    Functional groups are not yet supported
+
+    Messages are supported, see get_message() and get_message(), but are
+    optional: interchange segments can be accessed without going through
+    messages.
+
+    https://www.stylusstudio.com/edifact/40100/UNB_.htm
+    https://www.stylusstudio.com/edifact/40100/UNZ_.htm
+    """
+    def __init__(
+            self,
+            sender: str,
+            recipient: str,
+            control_reference: str,
+            syntax_identifier: Tuple[str, int],
+            delimiters: Characters = Characters(),
+            timestamp: datetime.datetime = None,
+            *args, **kwargs
+    ):
+        super().__init__(*args, **kwargs)
+        self.sender = sender
+        self.recipient = recipient
+        self.control_reference = control_reference
+        self.syntax_identifier = syntax_identifier
+        self.delimiters = delimiters
+        self.timestamp = timestamp or datetime.datetime.now()
+
+    def get_header_segment(self) -> Segment:
+        return Segment(
+            "UNB",
+            [str(i) for i in self.syntax_identifier],
+            self.sender,
+            self.recipient,
+            [f'{self.timestamp:%y%m%d}', f'{self.timestamp:%H%M}'],
+            self.control_reference,
+            *self.extra_header_elements,
+        )
+
+    def get_footer_segment(self) -> Segment:
+        return Segment(
+            "UNZ",
+            str(len(self.segments)),
+            self.control_reference,
+        )
+
+    def get_messages(self) -> List[Message]:
+        message = None
+        for segment in self.segments:
+            if segment.tag == 'UNH':
+                if not message:
+                    message = Message(segment.elements[0], segment.elements[1])
+                else:
+                    raise EDISyntaxError(
+                        f"Missing UNT segment before new UNH: {segment}"
+                    )
+            elif segment.tag == 'UNT':
+                if message:
+                    yield message
+                else:
+                    raise EDISyntaxError(
+                        f'UNT segment without matching UNH: "{segment}"'
+                    )
+            else:
+                if message:
+                    message.add_segment(segment)
+
+    def add_message(self, message: Message) -> "Interchange":
+        segments = (
+            [message.get_header_segment()]
+            + message.segments
+            + [message.get_footer_segment()]
+        )
+        self.add_segments(i for i in segments if i is not None)
+        return self
+
+    @classmethod
+    def from_segments(
+        cls, segments: list or collections.Iterable
+    ) -> "Interchange":
+        segments = iter(segments)
+
+        first_segment = next(segments)
+        if first_segment.tag == 'UNA':
+            unb = next(segments)
+        elif first_segment.tag == 'UNB':
+            unb = first_segment
+        else:
+            raise EDISyntaxError('An interchange must start with UNB or UNA and UNB')
+        # Loosy syntax check :
+        if len(unb.elements) < 4:
+            raise EDISyntaxError('Missing elements in UNB header')
+
+        datetime_str = '-'.join(unb.elements[3])
+        timestamp = datetime.datetime.strptime(datetime_str, '%y%m%d-%H%M')
+        interchange = Interchange(
+            syntax_identifier=unb.elements[0],
+            sender=unb.elements[1],
+            recipient=unb.elements[2],
+            timestamp=timestamp,
+            control_reference=unb.elements[4],
+        )
+
+        if first_segment.tag == 'UNA':
+            interchange.has_una_segment = True
+            interchange.characters = Characters.from_str(first_segment.elements[0])
+
+        return interchange.add_segments(
+            segment for segment in segments if segment.tag != 'UNZ'
+        )