+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog for PYDIFACT
 
 
+[0.2] - UNRELEASED
+- BREAKING CHANGE : Remove SegmentCollection
+
+[0.1] - UNRELEASED
+- Deprecate SegmentCollection
+- Add Interchange and Message semantic and featured classes
 
 [0.0.5] - 2020-04-29
 - corrected pypi details